Fixed scroll speed.

+++ ios/src/conflux/synergy/Protocol.h
@@ -60,7 +60,7 @@ typedef enum
 
 - (void)dmup:(const CFXMouseButton)whichButton; // moves given mouse button up (btn release)
 
-- (void)dmwm:(int)x andWith:(int)y; // mouse wheel move
+- (void)dmwm:(int16_t)x andWith:(int16_t)y; // mouse wheel move
 
 - (void)dkdn:(UInt16)key;
